+++ Content.Omu.Common/Construction/BigMachineBeingBuiltComponent.cs
@@ -1,8 +1,4 @@
-using Robust.Shared.GameStates;
-using Robust.Shared.Prototypes;
-
-namespace Content.Shared._Starlight.Power.BluespaceHarvester;
-
+namespace Content.Omu.Common.Construction;
 
 /// <summary>
 /// Used to mark that a machine that is big i.e. 3x3 is being built
