+++ Content.Omu.Common/Construction/BigMachineBoardComponent.cs
@@ -1,8 +1,4 @@
-using Robust.Shared.GameStates;
-using Robust.Shared.Prototypes;
-
-namespace Content.Shared._Starlight.Power.BluespaceHarvester;
-
+namespace Content.Omu.Common.Construction;
 
 /// <summary>
 /// Used to mark that a machine board will build into a big machine i.e. a 3x3 instead of a 1x1
