+++ Content.Omu.Common/Construction/OmuConstructionBlockerEvent.cs
@@ -1,0 +1,5 @@
+namespace Content.Omu.Common.Construction
+{
+    [ByRefEvent]
+    public record struct BigBuildAttemptEvent(EntityUid Machine, EntityUid? User, bool Cancelled = false);
+}
