+++ Resources/Locale/en-US/_Omu/Construction/bigmachinebuild.ftl
@@ -1,0 +1,1 @@
+big-machine-build-no-room = There's no room to finish building this!
